Add solution for problem `B - Array Cancellation` Link: https://codeforces.com/contest/1405/problem/B Tags: constructive algorithms, implementation, *1000 Ref: https://codeforces.com/contest/1405/submission/96427973

+++ codeforces/1405/B.java
@@ -322,12 +322,12 @@ public class Main {
                     int n = fr.nextInt();
 
                     //List<Integer> list = new ArrayList<>();
-                    long[] arr = new long[n];
+                    int[] arr = new int[n];
                     long minCoins = 0;
                     for (int i=0;i<n;i++)
                     {
-                        arr[i] = fr.nextLong();
-                        long ele = arr[i];
+                        arr[i] = fr.nextInt();
+                        int ele = arr[i];
                         minCoins = Math.max(0L,minCoins+ele);
                     }
